requirements.txt
numpy==1.24.4
nuscenes_devkit==1.1.11
opencv_python==4.11.0.86
pyquaternion==0.9.9
rich==14.0.0
scikit_learn==1.3.2
scipy==1.15.3
tqdm==4.67.1
ultralytics==8.3.131
